添加了异步分析接口、消息推送功能

+++ src/main/java/com/yupi/springbootinit/model/vo/BiResponse.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class BiResponse {
 
-    private int chartId;
+    private Long chartId;
 
     private String genChart;
